updates to form

+++ db/migrate/20180924233314_create_flights.rb
@@ -6,21 +6,21 @@ class CreateFlights < ActiveRecord::Migration[5.2]
       t.string :identification
       t.string :arrival
       t.string :departure
-      t.integer :airplane_sel
-      t.integer :airplane_mel
-      t.integer :dual_received
-      t.integer :pic_hours
-      t.integer :day_hours
-      t.integer :night_hours
-      t.integer :cross_country_hours
-      t.integer :actual_instrument_hours
-      t.integer :simulated_instrument
-      t.integer :ground_trainer
-      t.integer :num_instrument_approaches
-      t.integer :num_landings_day
-      t.integer :num_landings_night
-      t.integer :solo_hours
-      t.integer :total_duration
+      t.float :airplane_sel
+      t.float :airplane_mel
+      t.float :dual_received
+      t.float :pic_hours
+      t.float :day_hours
+      t.float :night_hours
+      t.float :cross_country_hours
+      t.float :actual_instrument_hours
+      t.float :simulated_instrument
+      t.float :ground_trainer
+      t.float :num_instrument_approaches
+      t.float :num_landings_day
+      t.float :num_landings_night
+      t.float :solo_hours
+      t.float :total_duration
       t.text :maneuvers
       t.text :comments
